init sdl

+++ examples/pong/src/constants.rs
@@ -1,8 +1,8 @@
 
-pub const WINDOW_WIDTH: u16 = 800;
-pub const WINDOW_HEIGHT: u16 = 600;
+pub const WINDOW_WIDTH: u32 = 800;
+pub const WINDOW_HEIGHT: u32 = 600;
 pub const BALL_GROUP:  u32 = 1000;
 pub const BALL_SPRITE_SHEET: usize = 0;
 pub const PADDLE_GROUP: u32 = 2000;
 pub const PADDLE_SPRITE_SHEET: usize = 1;
-pub const PADDLE_HALF_WIDTH: u16 = 16;
+pub const PADDLE_HALF_WIDTH: u32 = 16;
